refactor: authn -> internal/authn
