+++ internal/rpc/rpc.go
@@ -16,8 +16,8 @@ import (
 	"github.com/jpillora/backoff"
 	"golang.org/x/net/http2"
 
-	"github.com/TBD54566975/ftl/authn"
 	ftlv1 "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1"
+	"github.com/TBD54566975/ftl/internal/authn"
 	"github.com/TBD54566975/ftl/internal/log"
 )
